+++ src/library/DIPS.Mobile.UI/Components/Searching/Android/SearchBarHandler.cs
@@ -75,10 +75,8 @@ namespace DIPS.Mobile.UI.Components.Searching
         {
             SearchBarPropertyMapper.Add(nameof(SearchBar.CancelCommand), MapCancelCommand);
             SearchBarPropertyMapper.Add(nameof(SearchBar.CancelCommandParameter), MapCancelCommandParameter);
-            SearchBarPropertyMapper.Add(nameof(SearchBar.SearchCommand), MapSearchCommand);
             SearchBarPropertyMapper.Add(nameof(SearchBar.AndroidBusyBackgroundColor), MapAndroidBusyBackgroundColor);
             SearchBarPropertyMapper.Add(nameof(SearchBar.AndroidBusyColor), MapAndroidBusyColor);
-            SearchBarPropertyMapper.Add(nameof(SearchBar.SearchCommand), MapSearchCommand);
         }
 
         private static void MapAndroidBusyColor(SearchBarHandler handler, SearchBar searchBar)
@@ -110,8 +108,21 @@ namespace DIPS.Mobile.UI.Components.Searching
                     //Fixes this issue : https://github.com/dotnet/maui/issues/10823
                     mauiSearchView.MaxWidth = int.MaxValue;
                 }
-            }
 
+                InternalSearchBar.SearchCommand = new Command(() =>
+                {
+                    if (VirtualView.ShouldCloseKeyboardOnReturnKeyTapped)
+                    {
+                        // An ugly workaround to hide keyboard on Android
+                        InternalSearchBar.IsEnabled = false;
+                        InternalSearchBar.IsEnabled = true;
+                        UnFocus();
+                    }
+
+                    VirtualView.SearchCommand?.Execute(null);
+
+                });
+            }
             SubscribeToEvents();
         }
 
@@ -165,11 +176,6 @@ namespace DIPS.Mobile.UI.Components.Searching
         private static void MapCancelButtonTextColor(SearchBarHandler handler, SearchBar searchBar)
         {
             handler.CancelButton.TextColor = searchBar.CancelButtonTextColor;
-        }
-
-        private static void MapSearchCommand(SearchBarHandler handler, SearchBar searchBar)
-        {
-            handler.InternalSearchBar.SearchCommand = searchBar.SearchCommand;
         }
 
         private static void MapHasBusyIndication(SearchBarHandler handler, SearchBar searchBar)
@@ -264,6 +270,17 @@ namespace DIPS.Mobile.UI.Components.Searching
         private static void MapText(SearchBarHandler handler, SearchBar searchBar)
         {
             handler.InternalSearchBar.Text = searchBar.Text;
+        }
+        
+        private static void MapReturnKeyType(SearchBarHandler handler, SearchBar searchBar)
+        {
+            if (handler.InternalSearchBar.Handler.PlatformView is MauiSearchView mauiSearchView)
+            {
+                mauiSearchView.ImeOptions = (int)(searchBar.ReturnKeyType == SearchBarReturnKeyType.Done
+                    ? ImeAction.Done
+                    : ImeAction.Search);
+            }
+
         }
 
         public partial void Focus()