+++ src/library/DIPS.Mobile.UI/Components/Searching/SearchBarHandler.cs
@@ -21,8 +21,9 @@ internal partial class SearchBarHandler
         [nameof(SearchBar.PlaceholderColor)] = MapPlaceholderColor,
         [nameof(SearchBar.Text)] = MapText,
         [nameof(SearchBar.CancelButtonTextColor)] = MapCancelButtonTextColor,
+        [nameof(SearchBar.ReturnKeyType)] = MapReturnKeyType
     };
-    
+
     private void OnInternalSearchBarFocused(object? sender, FocusEventArgs e)
     {
         VirtualView.SendFocused();